switched to float32 by default

+++ cca_zoo/data/utils.py
@@ -27,5 +27,5 @@ class CCA_Dataset(Dataset):
 
     def __getitem__(self, idx):
         label = self.labels[idx]
-        views = [view[idx] for view in self.views]
+        views = [view[idx].astype(np.float32) for view in self.views]
         return tuple(views), label
